+++ sedate/__init__.py
@@ -17,7 +17,7 @@ it might very well make sense to use a datetime wrapper library.
 import operator
 import pytz
 
-from calendar import weekday, monthrange
+from calendar import monthrange
 from datetime import datetime, time, timedelta
 
 
@@ -33,6 +33,7 @@ if TYPE_CHECKING:
     from .types import DateOrDatetime
     from .types import Direction
     from .types import TDateOrDatetime
+    from .types import TzInfo
     from .types import TzInfoOrName
 
 __version__ = '0.3.0'
@@ -45,13 +46,16 @@ class NotTimezoneAware(Exception):
     pass
 
 
-def ensure_timezone(timezone: 'TzInfoOrName') -> pytz.BaseTzInfo:
+def ensure_timezone(timezone: 'TzInfoOrName') -> 'TzInfo':
     """ Make sure the given timezone is a pytz timezone, not just a string. """
 
     if isinstance(timezone, str):
         return pytz.timezone(timezone)
 
-    return timezone
+    # NOTE: We make our lives slightly easier by accepting BaseTzInfo
+    #       but returning the more specific classes, since the function
+    #       signatures for normalize/localize differ from the base class
+    return timezone  # type:ignore[return-value]
 
 
 def as_datetime(value: 'DateLike') -> datetime:
@@ -78,10 +82,17 @@ def standardize_date(date: datetime, timezone: 'TzInfoOrName') -> datetime:
     if date.tzinfo is None:
         date = replace_timezone(date, timezone)
 
-    return to_timezone(date, 'UTC')
+    return to_timezone(date, pytz.UTC)
 
 
-def replace_timezone(date: datetime, timezone: 'TzInfoOrName') -> datetime:
+def replace_timezone(
+    date: datetime,
+    timezone: 'TzInfoOrName',
+    *,
+    is_dst: bool = False,
+    raise_non_existent: bool = False,
+    raise_ambiguous: bool = False,
+) -> datetime:
     """ Takes the given date and replaces the timezone with the given timezone.
 
     No conversion is done in this method, it's simply a safe way to do the
@@ -92,13 +103,37 @@ def replace_timezone(date: datetime, timezone: 'TzInfoOrName') -> datetime:
         date.replace(tzinfo=timezone('Europe/Zurich'))
 
         # do this:
-        calendar.replace_timezone(date, 'Europe/Zurich')
+        sedate.replace_timezone(date, 'Europe/Zurich')
+
+    By default this will pick standard time for ambiguous times but just
+    as with DstTzInfo.normalize you can use `is_dst` to either pick the
+    other.
+
+    If you want to detect ambiguous/missing datetimes in the given timezone
+    you may optionally raise `pytz.NonExistentTimeError` or
+    `pytz.AmbiguousTimeError`
 
     """
 
     timezone = ensure_timezone(timezone)
+    naive = date.replace(tzinfo=None)
+    may_raise = raise_non_existent or raise_ambiguous
+    try:
+        localized = timezone.localize(
+            naive,
+            is_dst=None if may_raise else is_dst
+        )
+    except pytz.NonExistentTimeError as error:
+        if raise_non_existent:
+            raise error
 
-    return timezone.normalize(timezone.localize(date.replace(tzinfo=None)))
+        localized = timezone.localize(naive, is_dst=is_dst)
+    except pytz.AmbiguousTimeError as error:
+        if raise_ambiguous:
+            raise error
+        localized = timezone.localize(naive, is_dst=is_dst)
+
+    return timezone.normalize(localized)
 
 
 def to_timezone(date: datetime, timezone: 'TzInfoOrName') -> datetime:
@@ -117,7 +152,7 @@ def to_timezone(date: datetime, timezone: 'TzInfoOrName') -> datetime:
 
 def utcnow() -> datetime:
     """ Returns a timezone-aware datetime.utcnow(). """
-    return replace_timezone(datetime.utcnow(), 'UTC')
+    return replace_timezone(datetime.utcnow(), pytz.UTC)
 
 
 def is_whole_day(
@@ -197,6 +232,7 @@ def count_overlaps(
     number of times start/end overlaps with any of the dates.
 
     """
+
     count = 0
 
     for otherstart, otherend in dates:
@@ -222,6 +258,7 @@ def align_date_to_day(
     2012-1-24 10:00 up   -> 2012-1-24 23:59:59'999999
 
     """
+
     assert direction in ('up', 'down')
 
     aligned = (0, 0, 0, 0) if direction == 'down' else (23, 59, 59, 999999)
@@ -258,7 +295,8 @@ def align_range_to_day(
     on the given timezone.
 
     """
-    assert start <= end, "{} - {} is an invalid range".format(start, end)
+    if start > end:
+        raise ValueError(f'{start} - {end} is an invalid range')
 
     return (
         align_date_to_day(start, timezone, 'down'),
@@ -279,11 +317,9 @@ def align_date_to_week(
     date = align_date_to_day(date, timezone, direction)
 
     if direction == 'down':
-        return date - timedelta(
-            days=weekday(date.year, date.month, date.day))
+        return date - timedelta(days=date.weekday())
     else:
-        return date + timedelta(
-            days=6 - weekday(date.year, date.month, date.day))
+        return date + timedelta(days=6 - date.weekday())
 
 
 def align_range_to_week(
@@ -330,20 +366,99 @@ def align_range_to_month(
     )
 
 
+def offset_date(
+    date: 'TDateOrDatetime',
+    delta: timedelta,
+    *,
+    is_dst: bool = False,
+    raise_non_existent: bool = False,
+    raise_ambiguous: bool = False
+) -> 'TDateOrDatetime':
+    """ For date and most datetimes it will be the same as adding the
+    the date and delta naively, but for datetimes with a DstTzInfo it
+    will make sure a DST <-> ST transition won't shift the time by an
+    hour backwards or forwards.
+
+    Refer to :func:`replace_timezone` for behaviour of `is_dst`
+    """
+    if not isinstance(date, datetime):
+        return date + delta
+
+    if not isinstance(date.tzinfo, pytz.tzinfo.DstTzInfo):
+        return date + delta
+
+    tzinfo = date.tzinfo
+    naive_end = date.replace(tzinfo=None) + delta
+    return replace_timezone(
+        naive_end,
+        tzinfo,
+        is_dst=is_dst,
+        raise_non_existent=raise_non_existent,
+        raise_ambiguous=raise_ambiguous
+    )
+
+
 def get_date_range(
     day: datetime,
     start_time: time,
-    end_time: time
+    end_time: time,
+    *,
+    is_dst: bool = False,
+    raise_non_existent: bool = False,
+    raise_ambiguous: bool = False
 ) -> Tuple[datetime, datetime]:
-    """ Returns the date-range of a date a start and an end time. """
+    """ Returns the date-range of a date, a start and an end time.
 
-    start = datetime.combine(day.date(), start_time).replace(tzinfo=day.tzinfo)
-    end = datetime.combine(day.date(), end_time).replace(tzinfo=day.tzinfo)
+    For timezones with daylight savings this might return a range
+    that goes backwards on a ST -> DST transition::
+
+    2.30 -> 3.00 will result in 2.30 ST -> 3.00 DST
+
+    which is the same as 3.30 DST -> 3.00 DST
+
+    If you wish to properly handle this case and ambiguous times
+    then you will need to use `raise_non_existent=True` in order
+    to be able to respond to `pytz.AmbiguousTimeError`.
+
+    """
+
+    date = day.date()
+    tzinfo = day.tzinfo
+
+    start = datetime.combine(date, start_time)
+    end = datetime.combine(date, end_time)
+
+    if isinstance(tzinfo, pytz.tzinfo.DstTzInfo):
+        # we need to be more careful about DST <-> ST transitions
+        start = replace_timezone(
+            start,
+            tzinfo,
+            is_dst=is_dst,
+            raise_ambiguous=raise_ambiguous,
+            raise_non_existent=raise_non_existent
+        )
+        end = replace_timezone(
+            end,
+            tzinfo,
+            is_dst=is_dst,
+            raise_ambiguous=raise_ambiguous,
+            raise_non_existent=raise_non_existent
+        )
+    else:
+        # in this case we can just replace it back in
+        start = start.replace(tzinfo=tzinfo)
+        end = end.replace(tzinfo=tzinfo)
 
     # since the user can only one date with separate times it is assumed
     # that an end before a start is meant for the following day
     if end < start:
-        end += timedelta(days=1)
+        end = offset_date(
+            end,
+            timedelta(days=1),
+            is_dst=is_dst,
+            raise_non_existent=raise_non_existent,
+            raise_ambiguous=raise_ambiguous
+        )
 
     return start, end
 
@@ -363,15 +478,49 @@ def parse_time(timestring: str) -> time:
 
 def dtrange(
     start: 'TDateOrDatetime',
-    end: 'TDateOrDatetime',
-    step: timedelta = timedelta(days=1)
+    end: 'DateOrDatetime',
+    step: timedelta = timedelta(days=1),
+    *,
+    skip_missing: bool = False
 ) -> Iterator['TDateOrDatetime']:
     """ Yields dates between start and end (inclusive) using the given step
     size. The step size may be negative iff end < start.
 
     The type of start/end is kept (datetime => datetime, date => date).
 
+
+    The below only applies to timezones with daylight savings::
+
+    For ambiguous time periods the overlapping DST hour will be skipped.
+
+    For non-existent time periods we snap to the previous hour, so the same
+    hour will potentially be returned twice (once in ST and once in DST).
+    You can instead skip the missing hour by passing `skip_missing=True`
+
     """
+
+    # TODO: We might want to be able to optionally include the duplicate
+    #       hour on DST -> ST transition days. But that will get really
+    #       messy for steps smaller than an hour, if we want to preserve
+    #       the absolute order of the datetimes. It would probably be
+    #       easier/smarter to use dateutil at that point
+    tzinfo = None
+    if isinstance(start, datetime):
+        if isinstance(start.tzinfo, pytz.tzinfo.DstTzInfo):
+            # we want the underspecified version that doesn't know yet
+            # whether
+            tzinfo = start.tzinfo
+
+            # we convert to a tz-naive datetimes
+            start = start.replace(tzinfo=None)
+            if isinstance(end, datetime):
+                # before we convert the end to naive we need to make
+                # sure they're not in completely different timezones
+                assert isinstance(end.tzinfo, pytz.BaseTzInfo)
+                if end.tzinfo.zone != tzinfo.zone:
+                    end = to_timezone(end, tzinfo)
+
+                end = end.replace(tzinfo=None)
 
     if start <= end:
         remaining = operator.le
@@ -381,17 +530,27 @@ def dtrange(
         if step.total_seconds() > 0:
             step = timedelta(seconds=step.total_seconds() * -1)
 
-    while remaining(start, end):
-        yield start
-
+    def date_iter() -> Iterator['TDateOrDatetime']:
         # dates are immutable, so no copy is needed
-        start += step
+        current = start
+        while remaining(current, end):
+            yield current
+            current += step
 
+    if not tzinfo:
+        # we don't need to convert values back to the source timezone
+        yield from date_iter()
+        return
 
-def weeknumber(day: 'DateOrDatetime') -> int:
-    """ The weeknumber of the given date/dateime as defined by ISO 8601. """
-    # FIXME: This is super inefficient
-    return int(day.strftime('%V'))
+    # back-convert each element to original timezone
+    for date in date_iter():
+        assert isinstance(date, datetime)
+        try:
+            yield replace_timezone(
+                date, tzinfo, raise_non_existent=skip_missing
+            )
+        except pytz.NonExistentTimeError:
+            pass
 
 
 def weekrange(
@@ -406,26 +565,30 @@ def weekrange(
     if start and end span multiple weeks a start/end pair for each week is
     returned (start being monday, end being sunday).
 
-    Like :func:`dtrange` this function works with both datetime and date.
-    Unlike :func:`dtrange` this function does not work backwards (start > end).
+    Like :func:`dtrange` this function works with both datetime and date
+    and works backwards. In the backwards case the start/end pairs will be
+    backwards as well.
 
     """
 
-    if start > end:
-        # TODO: We can probably make this work backwards as well...
-        raise ValueError(f'{start} - {end} is an invalid range')
+    # we start by aligning the start to the end of the 1st week
+    # we then iterate over the end dates of the week
+    if start <= end:
+        day_step = timedelta(days=1)
+        week_step = timedelta(days=7)
+        aligned_start = start + timedelta(days=6 - start.weekday())
+    else:
+        # here the start and end are backwards
+        day_step = timedelta(days=-1)
+        week_step = timedelta(days=-7)
+        aligned_start = start - timedelta(days=start.weekday())
 
-    s = e = start
-    last_week = weeknumber(start)
+    s = start
+    for e in dtrange(aligned_start, end, week_step):
+        yield s, e
+        s = offset_date(e, day_step)
 
-    # FIXME: doing a step of 1 day is really dumb and makes us do
-    #        much more work than we need to
-    for day in dtrange(start, end):
-        if last_week != weeknumber(day):
-            yield s, e
-            s = e = day
-            last_week = weeknumber(s)
-        else:
-            e = day
-
-    yield s, e
+    # in the rare case that the end is perfectly aligned we will
+    # already have yielded it, so only yield it if we need to
+    if e != end:
+        yield s, end